Flatten Word part of Arg
